+++ javax/src/test/java/me/panpf/javax/test/security/DigestxTest.java
@@ -48,11 +48,28 @@ public class DigestxTest {
     @Test
     public void testInputStreamDigest() throws IOException {
         Assert.assertEquals(Digestx.getDigest(IOStreamx.byteInputStream(TEST_TEXT), "MD2"), TEXT_MD2);
+        Assert.assertEquals(Digestx.getDigestOrEmpty(IOStreamx.byteInputStream(TEST_TEXT), "MD2"), TEXT_MD2);
+        Assert.assertEquals(Digestx.getDigestOrNull(IOStreamx.byteInputStream(TEST_TEXT), "MD2"), TEXT_MD2);
+
         Assert.assertEquals(Digestx.getMD5(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_MD5);
+        Assert.assertEquals(Digestx.getMD5OrEmpty(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_MD5);
+        Assert.assertEquals(Digestx.getMD5OrNull(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_MD5);
+
         Assert.assertEquals(Digestx.getMD5_16(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_MD5_16);
+        Assert.assertEquals(Digestx.getMD5_16OrEmpty(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_MD5_16);
+        Assert.assertEquals(Digestx.getMD5_16OrNull(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_MD5_16);
+
         Assert.assertEquals(Digestx.getSHA1(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA1);
+        Assert.assertEquals(Digestx.getSHA1OrEmpty(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA1);
+        Assert.assertEquals(Digestx.getSHA1OrNull(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA1);
+
         Assert.assertEquals(Digestx.getSHA256(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA256);
+        Assert.assertEquals(Digestx.getSHA256OrEmpty(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA256);
+        Assert.assertEquals(Digestx.getSHA256OrNull(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA256);
+
         Assert.assertEquals(Digestx.getSHA512(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA512);
+        Assert.assertEquals(Digestx.getSHA512OrEmpty(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA512);
+        Assert.assertEquals(Digestx.getSHA512OrNull(IOStreamx.byteInputStream(TEST_TEXT)), TEXT_SHA512);
     }
 
     @Test
@@ -78,20 +95,40 @@ public class DigestxTest {
     @Test
     public void testFileDigest() throws IOException {
         File file = new File("/tmp/testDigest.tmp");
-        Filex.writeText(file, TEST_FILE_TEXT);
+        try {
+            Filex.writeText(file, TEST_FILE_TEXT);
 
-        Assert.assertEquals(Digestx.getDigest(file, "MD2"), FILE_MD2);
-        Assert.assertEquals(Digestx.getMD5(file), FILE_MD5);
-        Assert.assertEquals(Digestx.getMD5_16(file), FILE_MD5_16);
-        Assert.assertEquals(Digestx.getSHA1(file), FILE_SHA1);
-        Assert.assertEquals(Digestx.getSHA256(file), FILE_SHA256);
-        Assert.assertEquals(Digestx.getSHA512(file), FILE_SHA512);
+            Assert.assertEquals(Digestx.getDigest(file, "MD2"), FILE_MD2);
+            Assert.assertEquals(Digestx.getDigestOrEmpty(file, "MD2"), FILE_MD2);
+            Assert.assertEquals(Digestx.getDigestOrNull(file, "MD2"), FILE_MD2);
 
-        file.delete();
+            Assert.assertEquals(Digestx.getMD5(file), FILE_MD5);
+            Assert.assertEquals(Digestx.getMD5OrEmpty(file), FILE_MD5);
+            Assert.assertEquals(Digestx.getMD5OrNull(file), FILE_MD5);
+
+            Assert.assertEquals(Digestx.getMD5_16(file), FILE_MD5_16);
+            Assert.assertEquals(Digestx.getMD5_16OrEmpty(file), FILE_MD5_16);
+            Assert.assertEquals(Digestx.getMD5_16OrNull(file), FILE_MD5_16);
+
+            Assert.assertEquals(Digestx.getSHA1(file), FILE_SHA1);
+            Assert.assertEquals(Digestx.getSHA1OrEmpty(file), FILE_SHA1);
+            Assert.assertEquals(Digestx.getSHA1OrNull(file), FILE_SHA1);
+
+            Assert.assertEquals(Digestx.getSHA256(file), FILE_SHA256);
+            Assert.assertEquals(Digestx.getSHA256OrEmpty(file), FILE_SHA256);
+            Assert.assertEquals(Digestx.getSHA256OrNull(file), FILE_SHA256);
+
+            Assert.assertEquals(Digestx.getSHA512(file), FILE_SHA512);
+            Assert.assertEquals(Digestx.getSHA512OrEmpty(file), FILE_SHA512);
+            Assert.assertEquals(Digestx.getSHA512OrNull(file), FILE_SHA512);
+        } finally {
+            //noinspection ResultOfMethodCallIgnored
+            file.delete();
+        }
     }
 
     @Test
-    public void  testError() {
+    public void testError() {
         try {
             Assert.assertEquals(Digestx.getDigest(TEST_TEXT, "MD25"), TEXT_MD2);
             Assert.fail();